% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rerun_Human_liver_AUTD17.R
\docType{data}
\name{db_rerun_Human_Liver_AUTD17}
\alias{db_rerun_Human_Liver_AUTD17}
\title{db_rerun_Human_Liver_AUTD17}
\format{
A data frame with 6315 rows and 18 variables:
\describe{
  \item{\code{Peptide}}{character Peptide}
  \item{\code{X.10lgP}}{double X.10lgP}
  \item{\code{Mass}}{double Mass}
  \item{\code{Length}}{integer Peptide sequence}
  \item{\code{ppm}}{double ppm}
  \item{\code{m.z}}{double mass-to-charge}
  \item{\code{Z}}{integer charge}
  \item{\code{RT}}{double Retention Time}
  \item{\code{Area}}{double Area}
  \item{\code{Fraction}}{integer Fraction}
  \item{\code{Id}}{integer Id}
  \item{\code{Scan}}{character Scan}
  \item{\code{from.Chimera}}{character from.Chimera}
  \item{\code{Source.File}}{character Source.File}
  \item{\code{Accession}}{character Accession}
  \item{\code{PTM}}{character Post-Translational Modification}
  \item{\code{AScore}}{character AScore}
  \item{\code{Found.By}}{character Found.By}
}
}
\usage{
db_rerun_Human_Liver_AUTD17
}
\description{
database search results from the second run using PEAKS software,
using the raw file from HLA ligand ATLAS, Human Liver AutDonor 17
}
\details{
database search results from the second run using PEAKS software,
using the raw file from HLA ligand ATLAS, Human Liver AutDonor 17
}
\keyword{datasets}
