% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_checknetmhcpan_results.R
\name{export_checknetmhcpan_results}
\alias{export_checknetmhcpan_results}
\title{export_checknetmhcpan_results}
\usage{
export_checknetmhcpan_results(list_checknetMHCpan_results, export_dir)
}
\arguments{
\item{list_checknetMHCpan_results}{the results generated from running
checknetMHCpan()}

\item{export_dir}{the export directory where the results .csv files should
be exported.}
}
\value{
exports a folder containing three files
\enumerate{
           \item netMHCpan results in long format (the original output)(.csv file)
           \item netMHCpan results tidied (in wide format) so as to summarize
           the information per peptide (.tsv tab-separated file)
           \item the updated database search results which contain the categorizatiosn
           of the peptides found in common between the 2nd database search and
           the HybridFinder function (.csv file)}
}
\description{
this function allows to export the results generated from
checknetMHCpan()
}
\details{
In order to be able to have the checknetMHCpan() function results
exported, this function will come in handy. Please note that this function is
also part of the checknetMHCpan() function (if export_files is set to TRUE
and a valid export directory is indicated)
}
\examples{
\dontrun{
 export_checknetmhcpan_results(results_checknetMHCpan_Human_Liver_AUTD17, folder_Human_Liver_AUTD17)
}
}
