% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmhcpan_list_alleles.R
\docType{data}
\name{netmhcpan_list_alleles}
\alias{netmhcpan_list_alleles}
\title{netmhcpan_list_alleles}
\format{
A data frame with 1024 rows and 1 variables:
\describe{
  \item{\code{V1}}{character the alleles}
}
}
\usage{
netmhcpan_list_alleles
}
\description{
the list of alleles in the acceptable format for netMHCpan
}
\details{
the list of alleles in the acceptable format for netMHCpan
}
\keyword{datasets}
